Create heart_grid.py
In this exercise we rotated a grid like structure created through lists, the approach was to take the number of rows and the number of columns and then iterate again through each and every one of them but in a reverse order.

The main take away here are the challenges was to reduce the problem into a simple format and to understand it. if not looked at as an X&Y axis and you immediately dive into the for loops then it becomes near impossible to solve.